+++ lib/spack/spack/test/cmd/compiler.py
@@ -17,18 +17,6 @@ compiler = spack.main.SpackCommand('compiler')
 
 
 @pytest.fixture
-def no_compilers_yaml(mutable_config):
-    """Creates a temporary configuration without compilers.yaml"""
-
-    for scope, local_config in mutable_config.scopes.items():
-        compilers_yaml = os.path.join(
-            local_config.path, scope, 'compilers.yaml'
-        )
-        if os.path.exists(compilers_yaml):
-            os.remove(compilers_yaml)
-
-
-@pytest.fixture
 def mock_compiler_version():
     return '4.5.3'
